Day 13
Another slog of a day. Recursion was a distant 3rd year memory, and digging it up for R was unpleasant, especially with a not super clear explanation in Part 1.

Some good learning to be had in Part 2 though, now I know how to custom classes and implement custom sorting in R!